requirements.txt
colorzero==2.0
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.61.0
gpiozero==2.0.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.9.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smbus==1.1.post2
spicy==0.16.0
tk==0.1.0
tzdata==2025.2
